+++ lib/libsesame3bt-core/transport.cpp
@@ -20,12 +20,12 @@ SesameBLETransport::send_data(const std::byte* pkt, size_t pkt_size, bool is_cry
 	std::array<std::byte, 1 + FRAGMENT_SIZE> fragment;  // 1 for header
 	int pos = 0;
 	for (size_t remain = pkt_size; remain > 0;) {
-		fragment[0] = packet_header_t{
-		    pos == 0,
-		    remain > FRAGMENT_SIZE ? packet_kind_t::not_finished
-		    : is_crypted           ? packet_kind_t::encrypted
-		                           : packet_kind_t::plain,
-		    std::byte{0}}.value;
+		fragment[0] = packet_header_t{pos == 0,
+		                              remain > FRAGMENT_SIZE ? packet_kind_t::not_finished
+		                              : is_crypted           ? packet_kind_t::encrypted
+		                                                     : packet_kind_t::plain,
+		                              std::byte{0}}
+		                  .value;
 		size_t ssz = std::min(remain, FRAGMENT_SIZE);
 		std::copy(pkt + pos, pkt + pos + ssz, &fragment[1]);
 		if (!backend.write_to_tx(to_cptr(fragment), ssz + 1)) {
@@ -100,13 +100,13 @@ SesameBLETransport::reset() {
 	buffer.reset();
 }
 
-void
-SesameBLETransport::disconnect() {
-	backend.disconnect();
-	reset();
-}
+// void
+// SesameBLETransport::disconnect() {
+// 	backend.disconnect();
+// 	reset();
+// }
 
-bool
+result_t
 SesameBLETransport::send_notify(Sesame::op_code_t op_code,
                                 Sesame::item_code_t item_code,
                                 const std::byte* data,
@@ -121,14 +121,14 @@ SesameBLETransport::send_notify(Sesame::op_code_t op_code,
 		plain[1] = to_byte(item_code);
 		std::copy(data, data + data_size, &plain[2]);
 		if (!crypt.encrypt(plain, sizeof(plain), pkt, sizeof(pkt))) {
-			return false;
+			return result_t::crypt_failure;
 		}
 	} else {
 		pkt[0] = to_byte(op_code);
 		pkt[1] = to_byte(item_code);
 		std::copy(data, data + data_size, &pkt[2]);
 	}
-	return send_data(pkt, pkt_size, is_crypted);
+	return send_data(pkt, pkt_size, is_crypted) ? result_t::success : result_t::transport_failure;
 }
 
 }  // namespace libsesame3bt::core